trans_id on lane transaction-y tables should be INT, not TINYINT

+++ pos/is4c-nf/lib/models/trans/DTransactionsModel.php
@@ -68,7 +68,7 @@ class DTransactionsModel extends BasicModel
     'numflag'    => array('type'=>'INT','default'=>0),
     'charflag'    => array('type'=>'VARCHAR(2)','default'=>"''"),
     'card_no'    => array('type'=>'INT','index'=>True),
-    'trans_id'    => array('type'=>'TINYINT'),
+    'trans_id'    => array('type'=>'INT'),
     'pos_row_id' => array('type'=>'BIGINT UNSIGNED', 'increment'=>true, 'primary_key'=>true),
     );
